Remove documentation and component files
Deleted multiple markdown documentation files and SwiftUI component files related to strategies, implementation, and reusable UI components. Also updated the Xcode project configuration and modified NutritionTrackerView.swift, likely to reflect the removal of these components and documentation.
