--- NuTrackDemo03/00_BE_STRATEGY.md
@@ -1,47 +1,0 @@
-# SwiftData 總體遷移策略
-
-本文檔旨在記錄將 NuTrack App 從基於 JSON 的模擬資料，全面遷移至以 SwiftData 為核心的持久化資料庫的總體三步驟戰略。
-
-這個方案的核心思想是 **漸進式、低風險** 地完成遷移，確保每一步都是一個穩定且可驗證的里程碑，從而規避一次性重構可能帶來的風險。
-
----
-
-## 步驟一：【打地基】建立 SwiftData 模型與容器 (The Foundation)
-
-**目標**：在不影響任何現有 App 功能的前提下，建立起資料庫的「空殼」和「藍圖」。
-
-**具體工作**：
-1.  **定義模型 (Schema)**：建立 `UserProfile.swift` 和 `MealEntry.swift` 檔案，並在其中使用 `@Model` 關鍵字來定義我們的資料結構。
-2.  **建立容器 (Container)**：在 App 的主入口檔案 `NuTrackDemo03App.swift` 中，加入 `.modelContainer()` 修飾符，告訴 App 要根據我們定義的模型來準備資料庫。
-
-**完成時的狀態**：App 在啟動時，會在裝置上自動建立一個空的 SQLite 資料庫檔案。此時 App 的所有功能和畫面看起來都沒有任何改變，但我們已經為後續的資料儲存鋪好了穩固的地基。
-
----
-
-## 步驟二：【填資料】從 JSON 匯入種子資料 (The Seed)
-
-**目標**：將我們現有的模擬資料，一次性地匯入到上一步建立好的空資料庫中，使其成為持久化的資料。
-
-**具體工作**：
-1.  **編寫匯入邏輯**：建立一個一次性的資料匯入工具。
-2.  **讀取舊資料**：這個工具會讀取 `mock_users.json` 和 `mock_meals.json` 檔案的內容。
-3.  **寫入新資料庫**：將讀取到的 JSON 資料，轉換成 `UserProfile` 和 `MealEntry` 物件，並透過 SwiftData 的 `modelContext` 將它們插入到資料庫中。
-4.  **設定觸發時機**：這段邏輯通常只在 App 第一次啟動、且偵測到資料庫為空時執行一次。
-
-**完成時的狀態**：我們的 SwiftData 資料庫現在已經有了一份與舊的 JSON 完全一致的資料。App 的畫面和功能依然沒有改變，因為它們還在讀取舊的 JSON 檔案，但我們的「新水源」已經準備就緒。
-
----
-
-## 步驟三：【接活水】逐步重構視圖與邏輯 (The Switch)
-
-**目標**：讓 App 的所有部分，都從讀取舊的 `JSONDataManager`，切換為從 SwiftData 資料庫讀取資料。這是整個遷移工程的最後一步。
-
-**具體工作**：
-1.  **擇一改造**：選擇一個最簡單的視圖（View）開始，例如顯示今日食物日誌的 `TodayFoodLogView`。
-2.  **替換資料源**：
-    *   **移除** 該視圖中所有呼叫 `JSONDataManager` 的舊程式碼。
-    *   **換上** SwiftData 的 `@Query` 屬性包裝器，讓視圖可以直接、響應式地從資料庫中獲取資料。
-3.  **重複此過程**：對專案中每一個用到舊資料的視圖和元件，重複「移除舊程式碼、換上 @Query」的步驟。
-4.  **功成身退**：當整個專案再也沒有任何地方引用 `JSONDataManager` 和 `mock_meals.json` 時，就可以安全地將這些舊檔案刪除。
-
-**完成時的狀態**：一個功能完整、資料持久化、且完全由 SwiftData 驅動的現代化 App。整個遷移工程至此圓滿完成。
