--- NuTrackDemo03/DOCUMENTATION_GUIDELINES.md
@@ -1,52 +1,0 @@
-# 開發說明文件命名原則
-
-為了確保本專案的文件結構清晰、易於維護且具備擴充性，所有新增的計畫或說明文件都應遵循以下命名原則。
-
-## 命名格式
-
-所有文件應採用以下格式命名：
-
-**`[順序]_[領域]_[類型].md`**
-
----
-
-### 1. `[順序]` (Order)
-
-*   **用途**：控制檔案在目錄中的顯示順序，引導閱讀路徑。
-*   **格式**：使用兩位數編號，從 `00` 開始。
-*   **分組**：建議為不同的大型主題（如後端、前端）分配不同的數字區間，以形成邏輯分組。
-    *   `0x` 系列：用於後端 (Backend) 相關文件。
-    *   `1x` 系列：用於前端 (Frontend) 相關文件。
-    *   `2x` 系列：用於測試 (Testing) 相關文件... 以此類推。
-
-### 2. `[領域]` (Domain)
-
-*   **用途**：標識文件所屬的開發領域。
-*   **格式**：一個簡短的大寫縮寫。
-*   **已定義的領域**：
-    *   `BE`：**Backend**。泛指所有與資料處理、業務邏輯、SwiftData、服務相關的內容。
-    *   `FE`：**Frontend**。泛指所有與 UI、UX、視圖、元件、使用者互動相關的內容。
-
-### 3. `[類型]` (Type)
-
-*   **用途**：說明文件的具體性質。
-*   **格式**：一個簡潔的大寫單字。
-*   **建議的類型**：
-    *   `STRATEGY`：高層次的戰略、方向或總體規劃。
-    *   `IMPLEMENTATION`：具體的實作計畫、步驟或程式碼設計。
-    *   `NOTES`：會議記錄、靈感、或非正式的筆記。
-    *   `RFC`：(Request for Comments) 用於提出一個新想法，並徵求團隊意見的草案。
-
----
-
-### 範例
-
-一個結構良好的文件目錄看起來會像這樣：
-
-```
-.
-├── 00_BE_STRATEGY.md
-├── 01_BE_IMPLEMENTATION.md
-├── 10_FE_STRATEGY.md
-└── 11_FE_IMPLEMENTATION.md
-```
